Reorganize authentication and security packages
