+++ modules/fbs-identity-service/api/src/main/kotlin/de/thm/ii/fbs/fbs_identity_service/security/config/SecurityContextConfig.kt
@@ -1,4 +1,4 @@
-package de.thm.ii.fbs.fbs_identity_service.config
+package de.thm.ii.fbs.fbs_identity_service.security.config
 
 import org.springframework.context.annotation.Bean
 import org.springframework.context.annotation.Configuration
