+++ modules/fbs-identity-service/api/src/main/kotlin/de/thm/ii/fbs/fbs_identity_service/security/saml/SamlUser.kt
@@ -1,4 +1,4 @@
-package de.thm.ii.fbs.fbs_identity_service.model.auth
+package de.thm.ii.fbs.fbs_identity_service.security.saml
 
 data class SamlUser(
     val username: String,
