+++ modules/fbs-identity-service/api/src/test/kotlin/de/thm/ii/fbs/fbs_identity_service/security/saml/SamlAuthSuccessHandlerTest.kt
@@ -1,11 +1,10 @@
-package de.thm.ii.fbs.fbs_identity_service.config
+package de.thm.ii.fbs.fbs_identity_service.security.saml
 
 import de.thm.ii.fbs.fbs_identity_service.model.user.GlobalRole
 import de.thm.ii.fbs.fbs_identity_service.model.user.User
-import de.thm.ii.fbs.fbs_identity_service.security.local.IdentityUserPrincipal
-import de.thm.ii.fbs.fbs_identity_service.service.auth.saml.SamlLoginService
+import de.thm.ii.fbs.fbs_identity_service.security.principal.IdentityUserPrincipal
 import org.junit.jupiter.api.Test
-import org.mockito.Mockito.mock
+import org.mockito.Mockito
 import org.mockito.kotlin.any
 import org.mockito.kotlin.argumentCaptor
 import org.mockito.kotlin.eq
@@ -25,9 +24,9 @@ import kotlin.test.assertTrue
 
 class SamlAuthSuccessHandlerTest {
 
-    private val samlLoginService = mock<SamlLoginService>()
+    private val samlLoginService = Mockito.mock<SamlLoginService>()
 
-    private val securityContextRepository = mock<SecurityContextRepository>()
+    private val securityContextRepository = Mockito.mock<SecurityContextRepository>()
 
     private val handler = SamlAuthSuccessHandler(
         samlLoginService = samlLoginService,
@@ -50,7 +49,7 @@ class SamlAuthSuccessHandlerTest {
             )
         )
 
-        val authentication = mock<Authentication>()
+        val authentication = Mockito.mock<Authentication>()
         whenever(authentication.principal).thenReturn(principal)
 
         val request = MockHttpServletRequest()
@@ -77,10 +76,10 @@ class SamlAuthSuccessHandlerTest {
             )
         )
 
-        val samlAuthentication = mock<Authentication>()
+        val samlAuthentication = Mockito.mock<Authentication>()
         whenever(samlAuthentication.principal).thenReturn(principal)
 
-        val user = mock<User>()
+        val user = Mockito.mock<User>()
         whenever(user.id).thenReturn(1L)
         whenever(user.username).thenReturn("niklas")
         whenever(user.globalRole).thenReturn(GlobalRole.USER)
